finishing fitur laporan

+++ src/model/SchoolExpense.java
@@ -1,12 +1,13 @@
 package model;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.time.LocalDate;
 
 public class SchoolExpense {
     private int id;
     private int category_id;
-    private float jumlah;
+    private BigDecimal jumlah;
     private LocalDate tanggal_pengeluaran;
     private String bukti_transaksi;
     private String keterangan;
@@ -32,11 +33,11 @@ public class SchoolExpense {
         this.category_id = category_id;
     }
 
-    public float getJumlah() {
+    public BigDecimal getJumlah() {
         return jumlah;
     }
 
-    public void setjumlah(float jumlah) {
+    public void setjumlah(BigDecimal jumlah) {
         this.jumlah = jumlah;
     }
